moving copy operation to a separate functor + code clean-up
